Check strength of your Password.py
This Python project could be a terrific complement to your app creation skills, since it allows you to verify that your password is strong enough.

It achieves this by counting the characters in a given password, including letters, digits, special characters, and whitespace characters, and then assigning a score based on these findings. Thus, it's a fantastic additional resource for learning about functions, conditional expressions, and string formatting.

The Python standard library's string and getpass modules are also used by us. The.getpass() function enables us to conceal our password when we enter it, while this gives us access to the entire range of string characters to compare with the character composition of our password.